Move `validateMaxEvents` to reuse
