+++ src/blocks/EventQuery/hooks/validateMaxEvents.ts
