+++ src/blocks/EventTable/config.ts
@@ -8,9 +8,9 @@ import {
 } from '@payloadcms/richtext-lexical'
 import type { Block, Field, FilterOptionsProps } from 'payload'
 import { ButtonBlock } from '../Button/config'
+import { validateMaxEvents } from '../EventQuery/hooks/validateMaxEvents'
 import { GenericEmbedLexical } from '../GenericEmbed/config'
 import { MediaBlockLexical } from '../MediaBlock/config'
-import { validateMaxEvents } from './hooks/validateMaxEvents'
 
 const defaultStylingFields: Field[] = [
   { name: 'heading', type: 'text' },
